Quiz logic almost done, awaiting integration for includes

+++ networking/src/BinaryStream.String.cpp
@@ -3,7 +3,7 @@
 
 template<>
 void BinaryStream::Insert<std::string>(const std::string & string) {
-    const uint16_t string_size = string.size();
+    const uint64_t string_size = string.size();
     Insert(string_size);
     std::vector<std::byte> vec(string_size);
     std::transform(string.begin(), string.end(), vec.begin(), [] (char c) { return std::byte(c); });
@@ -12,7 +12,7 @@ void BinaryStream::Insert<std::string>(const std::string & string) {
 
 template<>
 size_t BinaryStream::Read<std::string>(std::string & string, const size_t offset) const {
-    uint16_t string_size = 0;
+    uint64_t string_size = 0;
     Read(string_size, offset);
     const size_t read_size = sizeof(string_size) + string_size;
     string.resize(string_size);
@@ -23,7 +23,7 @@ size_t BinaryStream::Read<std::string>(std::string & string, const size_t offset
 
 template<>
 size_t BinaryStream::Erase<std::string>(const size_t offset) {
-    uint16_t string_size = 0;
+    uint64_t string_size = 0;
     Read(string_size, offset);
     const size_t erase_size = sizeof(string_size) + string_size;
     if (m_data.size() < offset + erase_size) return 0;
